+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Allow transformers to work on top-level objects
 - Separate transformer to merge schemas containing `allOf`
+- Anomaly detection for different `ProductSchema` types
 
 ### Changed
 - Add separate types for `ProductSchema`